# backend/openapi.json
{
  "components": {
    "schemas": {
      "BoundaryCollection": {
        "description": "A GeoJSON FeatureCollection of boundary polygons.",
        "properties": {
          "features": {
            "items": {
              "$ref": "#/components/schemas/BoundaryFeature"
            },
            "title": "Features",
            "type": "array"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "features"
        ],
        "title": "BoundaryCollection",
        "type": "object"
      },
      "BoundaryFeature": {
        "description": "A GeoJSON feature for one boundary polygon.",
        "properties": {
          "geometry": {
            "additionalProperties": true,
            "title": "Geometry",
            "type": "object"
          },
          "properties": {
            "$ref": "#/components/schemas/BoundaryProperties"
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "properties",
          "geometry"
        ],
        "title": "BoundaryFeature",
        "type": "object"
      },
      "BoundaryProperties": {
        "description": "Properties of one boundary polygon.",
        "properties": {
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "feature_id": {
            "title": "Feature Id",
            "type": "string"
          },
          "name_en": {
            "title": "Name En",
            "type": "string"
          },
          "province": {
            "title": "Province",
            "type": "string"
          }
        },
        "required": [
          "feature_id",
          "district_id",
          "name_en",
          "province"
        ],
        "title": "BoundaryProperties",
        "type": "object"
      },
      "CheckResult": {
        "description": "Outcome of one dependency check.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "enum": [
              "pass",
              "fail"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "CheckResult",
        "type": "object"
      },
      "DailyForecast": {
        "description": "One day of forecast for one location. Missing values are None.",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "precipitation_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Mm"
          },
          "precipitation_probability_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Probability Pct"
          },
          "snowfall_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snowfall Cm"
          },
          "temperature_max_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Max C"
          },
          "temperature_min_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Min C"
          },
          "uv_index_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uv Index Max"
          },
          "weather_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Code"
          },
          "wind_gusts_max_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Gusts Max Kmh"
          },
          "wind_speed_max_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Max Kmh"
          }
        },
        "required": [
          "date",
          "temperature_max_c",
          "temperature_min_c",
          "precipitation_mm",
          "precipitation_probability_pct",
          "wind_speed_max_kmh",
          "wind_gusts_max_kmh",
          "weather_code",
          "snowfall_cm",
          "uv_index_max"
        ],
        "title": "DailyForecast",
        "type": "object"
      },
      "DistrictOut": {
        "description": "A district in the registry.",
        "properties": {
          "feature_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name_en": {
            "title": "Name En",
            "type": "string"
          },
          "name_ur": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Ur"
          },
          "province": {
            "title": "Province",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name_en",
          "name_ur",
          "province",
          "lat",
          "lon",
          "feature_id"
        ],
        "title": "DistrictOut",
        "type": "object"
      },
      "ForecastOut": {
        "description": "Daily forecast for a district, with where and when it came from.",
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/DailyForecast"
            },
            "title": "Days",
            "type": "array"
          },
          "district_id": {
            "title": "District Id",
            "type": "string"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "district_id",
          "source",
          "fetched_at",
          "stale",
          "days"
        ],
        "title": "ForecastOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Overall status plus the result of each check.",
        "properties": {
          "checks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CheckResult"
            },
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "enum": [
              "healthy",
              "unhealthy"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Early Warning System API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/districts": {
      "get": {
        "description": "List districts, optionally limited to one province.",
        "operationId": "list_districts_api_districts_get",
        "parameters": [
          {
            "description": "Province name, case-insensitive",
            "in": "query",
            "name": "province",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Province name, case-insensitive",
              "title": "Province"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DistrictOut"
                  },
                  "title": "Response List Districts Api Districts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Districts",
        "tags": [
          "districts"
        ]
      }
    },
    "/api/districts/geojson": {
      "get": {
        "description": "Boundary polygons as GeoJSON, each tagged with the district it belongs to.\n\nPolygons with no district in the registry are included with a null\n``district_id`` so the map has no holes.",
        "operationId": "district_boundaries_api_districts_geojson_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoundaryCollection"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "District Boundaries",
        "tags": [
          "districts"
        ]
      }
    },
    "/api/districts/{district_id}/forecast": {
      "get": {
        "description": "Latest daily forecast for a district, refreshed when it is too old.",
        "operationId": "district_forecast_api_districts__district_id__forecast_get",
        "parameters": [
          {
            "in": "path",
            "name": "district_id",
            "required": true,
            "schema": {
              "title": "District Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown district"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "description": "Forecast source unavailable and nothing stored"
          }
        },
        "summary": "District Forecast",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/health": {
      "get": {
        "description": "Report whether the service and its dependencies are usable.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    }
  }
}